requirements.txt
transformers
openai
pandas
python-dotenv
torch
tqdm